Removed crypt test trait and moved code inside blowfish test.
